backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
pillow==10.2.0
mediapipe==0.10.9
pymongo==4.6.1
numpy==1.26.0
pydantic==2.6.1
python-jose==3.3.0 